Update Irons_MacKenzie.txt
For stage 3 of our project, I worked alongside Julia and contributed when she sent her updated code. Julia was able to attend office hours to gain a better understanding of what was required and had more time spent developing the code. She would provide the updated code via email, and I was able to contribute a few times. Once the code became executable, I began working on the UML diagram. Sending emails with the contents that had been updated. Julia would inform me of changes that had been made to the code and then I would update the UML diagram. Conversing between text messages and emails we were both able to finalize the UML diagram to match the code. I then began creating the ReadMe.txt file to provide the user the instructions on how to compile and execute the code successfully. After sending that file to Julia we were able to converse the details to finalize that file. I then uploaded that into Github and created my own txt file. With the help and guidance of Julia, we were able to successfully upload all required documents into Github and submit the assignment.